requirements.txt
PyYAML==6.0.1
pyzabbix==1.2.1
urllib3==1.26.9
